Removed commented code and fixed the types
