+++ repos/fdbt-site/src/components/UserDataUploads.tsx
@@ -25,15 +25,7 @@ const UserDataUploadComponent = ({
     poundsOrPence,
     csrfToken,
     backHref,
-    buttonText,
-    dataSourceAttribute,
-    serviceList,
-    isFareZone,
-    clickedYes,
 }: UserDataUploadsProps): ReactElement => {
-    const seen: string[] = [];
-    const uniqueServiceLists =
-        serviceList?.filter((item) => (seen.includes(item.lineId) ? false : seen.push(item.lineId))) ?? [];
     return (
         <div className="govuk-grid-row">
             <div className="govuk-grid-column-two-thirds">
@@ -142,127 +134,6 @@ const UserDataUploadComponent = ({
                         )}
 
                         <input type="submit" value="Upload and continue" id="submit-button" className="govuk-button" />
-                        {isFareZone ? (
-                            <div>
-                                <div className="govuk-warning-text">
-                                    <span className="govuk-warning-text__icon" aria-hidden="true">
-                                        !
-                                    </span>
-                                    <strong className="govuk-warning-text__text">
-                                        <span className="govuk-warning-text__assistive">Warning</span>
-                                        If there are services exempt, you can omit them by selecting yes below and
-                                        selecting the services you want to omit.
-                                    </strong>
-                                </div>
-                                <div className="govuk-form-group">
-                                    <fieldset className="govuk-fieldset">
-                                        <legend className="govuk-fieldset__legend govuk-fieldset__legend--m">
-                                            <h2 className="govuk-fieldset__heading">
-                                                Are there services within this zone which are not included?
-                                            </h2>
-                                        </legend>
-                                        <FormElementWrapper
-                                            errorId="checkbox-0"
-                                            errorClass="govuk-form-group--error"
-                                            errors={errors}
-                                        >
-                                            <div className="govuk-radios" data-module="govuk-radios">
-                                                <div className="govuk-radios__item">
-                                                    <input
-                                                        className="govuk-radios__input"
-                                                        id="yes"
-                                                        name="exempt"
-                                                        type="radio"
-                                                        value="yes"
-                                                        data-aria-controls="conditional-yes"
-                                                        defaultChecked={clickedYes}
-                                                    />
-                                                    <label className="govuk-label govuk-radios__label" htmlFor="yes">
-                                                        Yes
-                                                    </label>
-                                                </div>
-                                                <div
-                                                    className="govuk-radios__conditional govuk-radios__conditional--hidden"
-                                                    id="conditional-yes"
-                                                >
-                                                    <div className="govuk-form-group">
-                                                        <fieldset className="govuk-fieldset">
-                                                            <input
-                                                                type="submit"
-                                                                name="selectAll"
-                                                                value={buttonText}
-                                                                id="select-all-button"
-                                                                className="govuk-button govuk-button--secondary"
-                                                            />
-                                                            <div className="govuk-checkboxes">
-                                                                {uniqueServiceLists.map((service, index) => {
-                                                                    const {
-                                                                        lineName,
-                                                                        lineId,
-                                                                        serviceCode,
-                                                                        description,
-                                                                        checked,
-                                                                        origin,
-                                                                        destination,
-                                                                    } = service;
-
-                                                                    const checkboxTitles =
-                                                                        dataSourceAttribute &&
-                                                                        dataSourceAttribute.source === 'tnds'
-                                                                            ? `${lineName} - ${description}`
-                                                                            : `${lineName} ${origin || 'N/A'} - ${
-                                                                                  destination || 'N/A'
-                                                                              }`;
-
-                                                                    const checkBoxValues = `${description}`;
-
-                                                                    return (
-                                                                        <div
-                                                                            className="govuk-checkboxes__item"
-                                                                            key={`checkbox-item-${lineName}`}
-                                                                        >
-                                                                            <input
-                                                                                className="govuk-checkboxes__input"
-                                                                                id={`checkbox-${index}`}
-                                                                                name={`${lineName}#${lineId}#${serviceCode}`}
-                                                                                type="checkbox"
-                                                                                value={checkBoxValues}
-                                                                                defaultChecked={checked}
-                                                                            />
-                                                                            <label
-                                                                                className="govuk-label govuk-checkboxes__label"
-                                                                                htmlFor={`checkbox-${index}`}
-                                                                            >
-                                                                                {checkboxTitles}
-                                                                            </label>
-                                                                        </div>
-                                                                    );
-                                                                })}
-                                                            </div>
-                                                        </fieldset>
-                                                    </div>
-                                                </div>
-                                                <div className="govuk-radios__item">
-                                                    <input
-                                                        className="govuk-radios__input"
-                                                        id="no"
-                                                        name="exempt"
-                                                        type="radio"
-                                                        value="no"
-                                                        defaultChecked={!clickedYes}
-                                                    />
-                                                    <label className="govuk-label govuk-radios__label" htmlFor="no">
-                                                        No
-                                                    </label>
-                                                </div>
-                                            </div>
-                                        </FormElementWrapper>
-                                    </fieldset>
-                                </div>
-                            </div>
-                        ) : (
-                            ''
-                        )}
                     </>
                 </CsrfForm>
             </div>
